pyproject.toml
[project]
name = "lexibrary"
version = "0.6.1"
description = "AI-friendly codebase indexer — creates .aindex files for LLM context navigation"
authors = [{ name = "Shannon Gray" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    # -- Phase 1: Foundation --
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pydantic>=2.9.0",
    "pathspec>=1.0.0,<2.0.0",
    # -- Token counting backends --
    "tiktoken>=0.9.0",
    "anthropic>=0.42.0",
    # -- Phase 4: LLM integration --
    "baml-py==0.218.1",
    # -- Config (YAML) --
    "PyYAML>=6.0.0,<7.0.0",
    # -- Dotenv loading --
    "python-dotenv>=1.0.0,<2.0.0",
    # -- Interactive prompts (wizard) --
    "questionary>=2.0.0",
    # -- AST parsing (tree-sitter) --
    "tree-sitter>=0.25.0,<0.26.0",
    "tree-sitter-python>=0.25.0,<0.26.0",
    "tree-sitter-javascript>=0.25.0,<0.26.0",
    "tree-sitter-typescript>=0.23.0,<0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.0",
    "mypy>=1.11.0",
    "pytest-asyncio>=0.25.0,<1.0.0",
]
# Phase 4: uncomment and pin when implementing LLM integration
# ollama = ["ollama>=0.6.0"]

[project.urls]
Homepage = "https://github.com/graysky-ai/Lexibrary"
Repository = "https://github.com/graysky-ai/Lexibrary"
Issues = "https://github.com/graysky-ai/Lexibrary/issues"

[project.scripts]
lexi = "lexibrary.cli:lexi_app"
lexictl = "lexibrary.cli:lexictl_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lexibrary"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
# `.env*` globs are excluded so pytest's rootdir walk does not `stat()` .env,
# which is denied by the Claude Code macOS sandbox and aborts collection.
addopts = "--ignore-glob=.env*"
markers = [
    "benchmark: performance benchmark tests; opt in with `-m benchmark`, skip with `-m 'not benchmark'`",
]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = "lexibrary.baml_client.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "baml_py.*"
ignore_missing_imports = true

# Legacy modules pending rework — uses retired v1 indexer types and old config attrs
[[tool.mypy.overrides]]
module = "lexibrary.crawler.engine"
ignore_errors = true

# tree-sitter grammar packages are C extensions without type stubs
[[tool.mypy.overrides]]
module = [
    "tree_sitter_python",
    "tree_sitter_typescript",
    "tree_sitter_javascript",
]
ignore_missing_imports = true

# questionary has no type stubs
[[tool.mypy.overrides]]
module = "questionary"
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["src/lexibrary/baml_client/"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "SIM",  # flake8-simplify
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
# Legacy crawler engine uses retired v1 indexer types (commented-out imports)
"src/lexibrary/crawler/engine.py" = ["F821"]
# AST parser fixtures: intentionally malformed/unconventional source for parser tests.
# They contain undefined names (F821), PascalCase function names (N802), and other
# constructs that exist solely to drive ClassEdgeSite extraction behavior.
"tests/test_ast_parser/fixtures/**/*.py" = ["F821", "N802"]
"tests/test_symbolgraph/fixtures/**/*.py" = ["F821", "N802"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.0",
    "mypy>=1.11.0",
    "pytest-asyncio>=0.25.0,<1.0.0",
    "openai>=2.20.0",
    "python-dotenv>=1.2.1",
    "types-pyyaml>=6.0.12.20250915",
]
